jacques_serveur/requirements.txt
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.7
django-braces==1.12.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.7.3
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
oauthlib==2.0.6
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
